==> code/notebooks/01_boxoffice.revenue.ipynb <==
# Grace Techau
## Box Office Revenue & Letterboxd Ratings Project 
### NOTEBOOK 1
### Importing & Cleaning Box Office Revenue Kaggle Data Set

# Import necessary packages
import pandas as pd 

## Import Kaggle data set of Box Office revnue data on movies from 2010 - "2010-2024 Movies Box Ofice Collection.csv"

box_office_revenue = pd.read_csv("2010-2024_Movies_Box_Ofice_Collection_raw.csv", encoding="utf-8")

# Display first 10 rows to make sure data frame loaded correctly 
print(box_office_revenue.shape)
display(box_office_revenue.head(10))

# create new "clean" data frame for cleaning 

# delete "Rank" column 
box_office_revenue_clean = box_office_revenue.drop('Rank', axis=1)

# rename columns with snake case & with proper names for later intergration
box_office_revenue_clean = box_office_revenue_clean.rename(columns={'Release Group': 'title', 'Worldwide':'worldwide_revenue', 'Domestic':'domestic_revenue','Domestic_percent':'domestic_percent', 'Foreign': 'foreign_revenue', 'Foreign_percent':'foreign_percent', 'year':'year'})

print("Clean Column Names")
print("-"*50)
print(box_office_revenue_clean.columns)

# Check data types 
print("Data Types of Dataframe before cleaning")
print("-"*50)
print(box_office_revenue_clean.dtypes)

# We need to change columns "worldwide_revenue", "domestic_revenue", "domestic percent" 
# "foreign_revenue" and "foreign_percent" all to numeric datatypes 

box_office_revenue_clean['worldwide_revenue'] = box_office_revenue_clean['worldwide_revenue'].str.replace(',','', regex=False)
box_office_revenue_clean['worldwide_revenue'] = pd.to_numeric(box_office_revenue_clean['worldwide_revenue'], errors='coerce')


box_office_revenue_clean['domestic_revenue'] = box_office_revenue_clean['domestic_revenue'].str.replace(',','', regex=False)
box_office_revenue_clean['domestic_revenue'] = pd.to_numeric(box_office_revenue_clean['domestic_revenue'], errors='coerce')

box_office_revenue_clean['foreign_revenue'] = box_office_revenue_clean['foreign_revenue'].str.replace(',','', regex=False)
box_office_revenue_clean['foreign_revenue'] = pd.to_numeric(box_office_revenue_clean['foreign_revenue'], errors='coerce')


box_office_revenue_clean['domestic_percent'] = box_office_revenue_clean['domestic_percent'].str.replace('%','', regex=False)
box_office_revenue_clean['domestic_percent'] = pd.to_numeric(box_office_revenue_clean['domestic_percent'], errors='coerce')

box_office_revenue_clean['foreign_percent'] = box_office_revenue_clean['foreign_percent'].str.replace('%','', regex=False)
box_office_revenue_clean['foreign_percent'] = pd.to_numeric(box_office_revenue_clean['foreign_percent'], errors='coerce')

# Strip any white space from title column
box_office_revenue_clean['title'] = box_office_revenue_clean['title'].str.strip()

print("Data Types of Dataframe after cleaning")
print("-"*50)
print(box_office_revenue_clean.dtypes)

# Conert the percentage fields into values between 0 and 1 & round to 2 decimal places 
box_office_revenue_clean['domestic_percent'] = ((box_office_revenue_clean['domestic_percent']) / 100).round(2)
box_office_revenue_clean['foreign_percent'] = ((box_office_revenue_clean['foreign_percent']) / 100).round(2)

#Check to see if they transformed correctly 
print("New percentage fields as values from 0 to 1")
display(box_office_revenue_clean[['domestic_percent', 'foreign_percent']].head(5))

# Change the three revenue columns to be scaled in millions and round to 2 decimal places 

box_office_revenue_clean['worldwide_revenue'] = (box_office_revenue_clean['worldwide_revenue'] / 1000000).round(2)
box_office_revenue_clean['domestic_revenue'] = (box_office_revenue_clean['domestic_revenue'] / 1000000).round(2)
box_office_revenue_clean['foreign_revenue'] = (box_office_revenue_clean['foreign_revenue'] / 1000000).round(2)

print("Revenue Columns Represented in Millions of U.S $")
print('-'*50)
display(box_office_revenue_clean[['worldwide_revenue', 'domestic_revenue', 'foreign_revenue']].head(5))

# Drop all records that are from the year 2010-2016 and 2020-2024 
years_to_drop = [2010, 2011, 2012, 2013, 2014, 2015, 2016, 2020, 2021, 2022, 2023]

rows_to_drop = box_office_revenue_clean[box_office_revenue_clean['year'].isin(years_to_drop)].index

box_office_revenue_clean = box_office_revenue_clean.drop(rows_to_drop)

# Check to see that only the years we want are included 
print("Years included after dropping 2010, 2020-2023")
print("-"*50)
print(box_office_revenue_clean['year'].unique())

# Some Movie Title names seemed to have replaced a "-" with a numerical 0 
# This was the way the data was in the Kaggle data set - nothing wrong with the encoding in reading the CSV file 

# Find all rows in title column where the movie name contains a '0'
print("List of Title Names with 0's in Them")
print("-"*50)
display(box_office_revenue_clean[box_office_revenue_clean['title'].str.contains('0', na=False)]['title'])
print("\n")

# There seems to be ~80 values that have had dashes (-) replaced with 0 
# We do not want to repalce ALL 0's with dashes 
# (eg. Movies with titles have years or numerical values with 0's in them)

# Define function to replace zeros  ('0') with dashes ('-')
def fix_dashes(title): 
    words = title.split() 
    for i, word in enumerate(words): 
        if word ==0: 
            words[i] == '-' 
        elif '0' in word and not word.replace('0', '').isdigit(): 
            words[i] = word.replace('0', '-')
    return ' '.join(words)

box_office_revenue_clean['title2'] = box_office_revenue_clean['title'].apply(fix_dashes)

print(f"Number of rows with 0's before function applied: {len(box_office_revenue_clean[box_office_revenue_clean['title'].str.contains('0', na=False)]['title'])}")
print(f"Number of rows with 0's after function applied: {len(box_office_revenue_clean[box_office_revenue_clean['title2'].str.contains('0', na=False)]['title'])}")
print("\n")

print("Title Names with 0's After Function Applied")
print("-"*50)
display(box_office_revenue_clean[box_office_revenue_clean['title2'].str.contains('0', na=False)]['title2'])
print("All dashes seemed to be fixed! The only 0's left are ones that are included 0's in years or numeric digits included in the title.")

# delete orginal title column 
box_office_revenue_clean.drop('title', axis=1, inplace=True)

#rename the modified title2 column to "title"
box_office_revenue_clean.rename(columns={'title2':'title'}, inplace=True)

print("Final Data Frame")
print("-"*50)
print(box_office_revenue_clean.shape)
display(box_office_revenue_clean)

# Save cleaned data frame to a distinctlly named CSV file 

box_office_revenue_clean.to_csv("box_office_revenue_2017.2019_clean.csv", header=True, index=False, encoding='utf-8')
