==> code/notebooks/02_letterboxd.2017.clean.ipynb <==
# Grace Techau
## Box Office Revenue & Letterboxd Ratings Project 
### NOTEBOOK 2 
### Cleaning Letterboxd Movie Data - 2017

#import all required packages 
import pandas as pd 
import re 

### Merge raw scraped data files for Letterboxd movies in 2017

The top 25% most popular Letterboxd movies in 2017 were scraped in four parts and were saved in seperate CSV files. The breakdown of pages scraped per file is below: 
- Pages 1 to 37.5 are captured in 'letterboxd_movie_data_2017_part1.1.csv'
- Pages 37.5 to 42.5 are captued in 'letterboxd_movie_data_2017_part1.2.csv'
- Pages 42.5 to 55.5 are captured in 'letterboxd_movie_data_2017_part1.3.csv'
- Pages 55.5 - 66 are captured in 'letterboxd_movie_data_2017_part1.4.csv''

These CSV files need to be merged into one complete 25% most popular 2017 Letterboxd movies Pandas data frame where all the data can be cleaned.  

# Read in 2017 scraping part 1 to a pandas data frame 

raw_2017_1 = pd.read_csv("letterboxd_movie_data_2017_raw_1.csv", encoding='utf-8')

print(f"Number of movies in 2017 part 1: {len(raw_2017_1)}")

# Read in 2017 scraping part 2 to a pandas data frame 

raw_2017_2 = pd.read_csv("letterboxd_movie_data_2017_raw_2.csv", encoding='utf-8')

print(f"Number of movies in 2017 part 2: {len(raw_2017_2)}")

# Read in 2017 scraping part 3 to a pandas data frame 

raw_2017_3 = pd.read_csv("letterboxd_movie_data_2017_raw_3.csv", encoding='utf-8')

print(f"Number of movies in 2017 part 3: {len(raw_2017_3)}")

# Read in 2017 scraping part 4 to a pandas data frame 

raw_2017_4 = pd.read_csv("letterboxd_movie_data_2017_raw_4.csv", encoding='utf-8')

print(f"Number of movies in 2017 part 4: {len(raw_2017_4)}")

# Merge the different data frames to one collective 2017 pandas data frame 

raw_movie_data_2017 = pd.concat([raw_2017_1, raw_2017_2, raw_2017_3, raw_2017_4], axis=0, ignore_index=True)

print(f"Total number of movies in 2017: {len(raw_movie_data_2017)}")
print("-"*50)
display(raw_movie_data_2017.head(5))

### Clean complete 2017 Letterboxd movie data set 

# Make a copy of the raw data frame for cleaning
clean_movie_data_2017 = raw_movie_data_2017.copy()

Since the 2017 Letterboxd movie data was scraped in batches and some of those batches ended in the middle of a page, there are duplicates present. I scraped the ending page of each batch as the starting page of the next batch to make sure all movies were scraped. Now those duplicates can be removed. 

duplicates_raw = raw_movie_data_2017[raw_movie_data_2017.duplicated(keep=False)]
print(f"Total duplicate records in raw 2017 data frame: {len(duplicates_raw)}")
print("-"*50)

clean_movie_data_2017 = clean_movie_data_2017.drop_duplicates(keep=False)

duplicates_clean = clean_movie_data_2017[clean_movie_data_2017.duplicated(keep=False)]
print(f"Total duplicate records in clean 2017 data frame: {len(duplicates_clean)}")

The analysis for this project is utilizing the rating data from Letterboxd, therefore any movies where the rating wasn't available are not relevant and can be dropped. 

no_rating_data = clean_movie_data_2017[(clean_movie_data_2017['average_rating'] == 'No average rating available') & (clean_movie_data_2017['number_ratings'] == 'No number of ratings available')].index

print("# rows where no average rating and number ratings was available was available")
print("-"*50)
display(len(clean_movie_data_2017.loc[no_rating_data]))

### Drop these rows 
clean_movie_data_2017 = clean_movie_data_2017.drop(no_rating_data)

print("\n")
print("length of the data frame after dropping rows with no rating data")
print("-"*50)
print(len(clean_movie_data_2017))

Remove text from 'length' column so that only the minutes numerical value is present. \
For example changing '164 mins More at IMDB TMDB' to '164'

# Keep only the numeric value from the string of text in the length column
clean_movie_data_2017.loc[:, 'length'] = clean_movie_data_2017['length'].str.replace(r'\D', '', regex=True)

print("length column before cleaning")
print("-"*50)
display(raw_movie_data_2017['length'].head(10))
print("\n")
print("length column after cleaning")
print("-"*50)
display(clean_movie_data_2017['length'].head(10))

Remove text from 'number_ratings' column to include only the numerical value for the number of ratings. \
For example changing 'Weighted average of 3.03 based on 288 ratings' to '288'

# using pandas .str.extract() method to keep only the numerical value after string of text 'based on'

clean_movie_data_2017['number_ratings'] = clean_movie_data_2017['number_ratings'].str.extract(r'based on ([\d,]+)').replace({',': ''}, regex=True)

print("number_ratings column before cleaning")
print("-"*50)
display(raw_movie_data_2017['number_ratings'].head(10))
print("\n")
print("number_ratings column after cleaning")
print("-"*50)
display(clean_movie_data_2017['number_ratings'].head(10))

Some movies have many genres (10+), for the scope of this project only keeping the first three genres is relevant. 

def clean_genres(genre_str):
    genres_list = genre_str.split(',')

    # remove duplicates while preserving order 
    unique_genres = list(dict.fromkeys(genre.strip() for genre in genres_list)) 
    
    # return only the first three unique genres 
    return ', '.join(unique_genres[:3])

clean_movie_data_2017['genres'] = clean_movie_data_2017['genres'].apply(clean_genres)

print("genres column before cleaning")
print("-"*50)
display(raw_movie_data_2017['genres'].head(10))
print("\n")

print("genres column after cleaning")
print("-"*50)
display(clean_movie_data_2017['genres'].head(10))

Correct all data types. 

print("data types before cleaning")
print("-"*50)
print(raw_movie_data_2017.dtypes)

# Remove any white space from all columns with object data type 
clean_movie_data_2017['title'] = clean_movie_data_2017['title'].str.strip()
clean_movie_data_2017['number_ratings'] = clean_movie_data_2017['number_ratings'].str.strip()
clean_movie_data_2017['average_rating'] = clean_movie_data_2017['average_rating'].str.strip()
clean_movie_data_2017['length'] = clean_movie_data_2017['length'].str.strip()
clean_movie_data_2017['genres'] = clean_movie_data_2017['genres'].str.strip()

# Convert neccessary columns to numeric data type 
clean_movie_data_2017['year'] = clean_movie_data_2017['year'].astype(int)
clean_movie_data_2017['number_ratings'] = clean_movie_data_2017['number_ratings'].astype(int)
clean_movie_data_2017['average_rating'] = clean_movie_data_2017['average_rating'].astype(float)

# There are some rows that have empty strings for length - convert these to 0's
clean_movie_data_2017['length'] = clean_movie_data_2017['length'].replace('', 0)
clean_movie_data_2017['length'] = clean_movie_data_2017['length'].astype(int)

print("\n")
print("data types after cleaning")
print("-"*50)
print(clean_movie_data_2017.dtypes)

### Final Comparison of Clean and Raw Data Frames 

print(f"Length of raw data frame: {len(raw_movie_data_2017)}")
print("-"*50)
display(raw_movie_data_2017.head(5))
print("\n")

print(f"Length of clean data frame: {len(clean_movie_data_2017)}")
print("-"*50)
display(clean_movie_data_2017.head(5))


### Save clean complete 2017 Letterboxd movie details for 25% most popular movies to CSV file 

clean_movie_data_2017.to_csv("letterboxd_movie_data_2017_clean.csv", header=True, index=False, encoding='utf-8')
