==> code/notebooks/03_letterboxd.2018.clean.ipynb <==
# Grace Techau
## Box Office Revenue & Letterboxd Ratings Project 
### NOTEBOOK 3
### Cleaning Letterboxd Movie Data - 2018

#import all required packages 
import pandas as pd 

### Merge raw scraped data files for Letterboxd movies in 2018

The top 25% most popular Letterboxd movies in 2018 were scraped in three parts and were saved in seperate CSV files. The breakdown of pages scraped per file is below: 
- Pages 1 to 35 are captured in letterboxd_movie_data_2018_raw_1.csv
- Pages 36 to 39 are captured in letterboxd_movie_data_2018_raw_2.csv
- Pages 39 to 71 are caputed in latterboxd_movie_data_2018_raw_3.csv

These CSV files need to be merged into one complete 25% most popular 2018 Letterboxd movies Pandas data frame where all the data can be cleaned.  

# Read in 2018 scraping part 1 to a pandas data frame 

raw_2018_1 = pd.read_csv("letterboxd_movie_data_2018_raw_1.csv", encoding='utf-8')

print(f"Number of movies in 2018 part 1: {len(raw_2018_1)}")

# Read in 2018 scraping part 2 to a pandas data frame 

raw_2018_2 = pd.read_csv("letterboxd_movie_data_2018_raw_2.csv", encoding='utf-8')

print(f"Number of movies in 2018 part 2: {len(raw_2018_2)}")

# Read in 2018 scraping part 3 to a pandas data frame 

raw_2018_3 = pd.read_csv("letterboxd_movie_data_2018_raw_3.csv", encoding='utf-8')

print(f"Number of movies in 2018 part 3: {len(raw_2018_3)}")

# Merge the different data frames to one collective 2018 pandas data frame 

raw_movie_data_2018 = pd.concat([raw_2018_1, raw_2018_2, raw_2018_3], axis=0, ignore_index=True)

print(f"Total number of movies in 2018: {len(raw_movie_data_2018)}")
print("-"*50)
display(raw_movie_data_2018.head(5))

### Clean complete 2018 Letterboxd movie data set 

# Make a copy of the raw data frame for cleaning
clean_movie_data_2018 = raw_movie_data_2018.copy()

Since the 2018 Letterboxd movie data was scraped in batches and some of those batches ended in the middle of a page, there are duplicates present. I scraped the ending page of each batch as the starting page of the next batch to make sure all movies were scraped. Now those duplicates can be removed. 

duplicates_raw = raw_movie_data_2018[raw_movie_data_2018.duplicated(keep=False)]
print(f"Total duplicate records in raw 2018 data frame: {len(duplicates_raw)}")
print("-"*50)

clean_movie_data_2018 = clean_movie_data_2018.drop_duplicates(keep=False)

duplicates_clean = clean_movie_data_2018[clean_movie_data_2018.duplicated(keep=False)]
print(f"Total duplicate records in clean 2018 data frame: {len(duplicates_clean)}")

The analysis for this project is utilizing the rating data from Letterboxd, therefore any movies where the rating wasn't available are not relevant and can be dropped. 

no_rating_data = clean_movie_data_2018[(clean_movie_data_2018['average_rating'] == 'No average rating available') & (clean_movie_data_2018['number_ratings'] == 'No number of ratings available')].index

print("# rows where no average rating and number ratings was available was available")
print("-"*50)
display(len(clean_movie_data_2018.loc[no_rating_data]))

### Drop these rows 
clean_movie_data_2018 = clean_movie_data_2018.drop(no_rating_data)

print("\n")
print("length of the data frame after dropping rows with no rating data")
print("-"*50)
print(len(clean_movie_data_2018))

Remove text from 'length' column so that only the minutes numerical value is present. \
For example changing '164 mins More at IMDB TMDB' to '164'

# Keep only the numeric value from the string of text in the length column
clean_movie_data_2018.loc[:, 'length'] = clean_movie_data_2018['length'].str.replace(r'\D', '', regex=True)

print("length column before cleaning")
print("-"*50)
display(raw_movie_data_2018['length'].head(10))
print("\n")
print("length column after cleaning")
print("-"*50)
display(clean_movie_data_2018['length'].head(10))

Remove text from 'number_ratings' column to include only the numerical value for the number of ratings. \
For example changing 'Weighted average of 3.03 based on 288 ratings' to '288'

# using pandas .str.extract() method to keep only the numerical value after string of text 'based on'

clean_movie_data_2018['number_ratings'] = clean_movie_data_2018['number_ratings'].str.extract(r'based on ([\d,]+)').replace({',': ''}, regex=True)

print("number_ratings column before cleaning")
print("-"*50)
display(raw_movie_data_2018['number_ratings'].head(10))
print("\n")
print("number_ratings column after cleaning")
print("-"*50)
display(clean_movie_data_2018['number_ratings'].head(10))

Some movies have many genres (10+), for the scope of this project only keeping the first three genres is relevant. 

def clean_genres(genre_str):
    genres_list = genre_str.split(',')

    # remove duplicates while preserving order 
    unique_genres = list(dict.fromkeys(genre.strip() for genre in genres_list)) 
    
    # return only the first three unique genres 
    return ', '.join(unique_genres[:3])

clean_movie_data_2018['genres'] = clean_movie_data_2018['genres'].apply(clean_genres)

print("genres column before cleaning")
print("-"*50)
display(raw_movie_data_2018['genres'].head(10))
print("\n")

print("genres column after cleaning")
print("-"*50)
display(clean_movie_data_2018['genres'].head(10))

Correct all data types. 

print("data types before cleaning")
print("-"*50)
print(raw_movie_data_2018.dtypes)

# Remove any white space from all columns with object data type 
clean_movie_data_2018['title'] = clean_movie_data_2018['title'].str.strip()
clean_movie_data_2018['number_ratings'] = clean_movie_data_2018['number_ratings'].str.strip()
clean_movie_data_2018['average_rating'] = clean_movie_data_2018['average_rating'].str.strip()
clean_movie_data_2018['length'] = clean_movie_data_2018['length'].str.strip()
clean_movie_data_2018['genres'] = clean_movie_data_2018['genres'].str.strip()

# Convert neccessary columns to numeric data type 
clean_movie_data_2018['year'] = clean_movie_data_2018['year'].astype(int)
clean_movie_data_2018['number_ratings'] = clean_movie_data_2018['number_ratings'].astype(int)
clean_movie_data_2018['average_rating'] = clean_movie_data_2018['average_rating'].astype(float)

# There are some rows that have empty strings for length - convert these to 0's
clean_movie_data_2018['length'] = clean_movie_data_2018['length'].replace('', 0)
clean_movie_data_2018['length'] = clean_movie_data_2018['length'].astype(int)

print("\n")
print("data types after cleaning")
print("-"*50)
print(clean_movie_data_2018.dtypes)

### Final Comparison of Clean and Raw Data Frames 

print(f"Length of raw data frame: {len(raw_movie_data_2018)}")
print("-"*50)
display(raw_movie_data_2018.head(5))
print("\n")

print(f"Length of clean data frame: {len(clean_movie_data_2018)}")
print("-"*50)
display(clean_movie_data_2018.head(5))


### Save clean complete 2018 Letterboxd movie details for 25% most popular movies to CSV file 

clean_movie_data_2018.to_csv("letterboxd_movie_data_2018_clean.csv", header=True, index=False, encoding='utf-8')
