==> code/notebooks/05_data.file.merging.ipynb <==
# Grace Techau
## Box Office Revenue & Letterboxd Ratings Project 
### NOTEBOOK 5
### Merging Clean Letterboxd and Box Office Revenue Files 

# import required libraries
import pandas as pd 

### Import full clean Letterboxd data files from years 2017, 2018 and 2019 to pandas data frames

letterboxd_2017 = pd.read_csv("letterboxd_movie_data_2017_clean.csv", encoding='utf-8')

print(f"Number of Letterboxd movies in 2017: {len(letterboxd_2017)}")

letterboxd_2018 = pd.read_csv("letterboxd_movie_data_2018_clean.csv", encoding='utf-8')

print(f"Number of Letterboxd movies in 2018: {len(letterboxd_2018)}")

letterboxd_2019 = pd.read_csv("letterboxd_movie_data_2019_clean.csv", encoding='utf-8')

print(f"Number of Letterboxd movies in 2019: {len(letterboxd_2019)}")

**Merge Letterboxd data frames from years 2017, 2018, and 2019 into one Letterboxd movie data frame**

letterboxd_merge = pd.concat([letterboxd_2017, letterboxd_2018, letterboxd_2019], axis=0, ignore_index=True)

print(f"Total number of movies scraped from Letterboxd: {len(letterboxd_merge)}")
print("-"*50)
display(letterboxd_merge.head(5))

Since the Letterboxd data was scraped this final merge of clean data across years could  have duplicates. I removed all duplicates. 

duplicates = letterboxd_merge[letterboxd_merge.duplicated(keep=False)]

print(f"total records: {len(duplicates)}")
print(f"duplicate records: {len(duplicates[duplicates.duplicated(keep="first")])}")

letterboxd_merge = letterboxd_merge.drop_duplicates()

**Save cleaned Letterboxd data from years 2017 to 2019 into one csv file** 

letterboxd_merge.to_csv('letterboxd_movie_data_2017.2019_clean.csv', header=True, index=False, encoding='utf-8')

## Re-open the full clean CSV file to another pandas data frame for merging with Box Office Data 

letterboxd_data = pd.read_csv('letterboxd_movie_data_2017.2019_clean.csv', encoding='utf-8')

### Import Box Office Revenue data file into pandas data frame

box_office = pd.read_csv("box_office_revenue_2017.2019_clean.csv", encoding='utf-8')

print(f"Number of movies from Kaggle Box Office Revenue: {len(box_office)}")
print("-"*50)
display(box_office.head(5))

## Merge Letterboxd and Box Office Revenue data frames

## Use an inner join to merge the data frames on 'title' and 'year'
## This will keep only matching rowsabs 

film_data = pd.merge(box_office, letterboxd_data, on=['title', 'year'], how='inner')

print(f"Number of movies in final data set: {len(film_data)}")
print("-"*50)
display(film_data.head(5))

## Final Cleaning to Full Data Set Before Analysis 

## Re-order the Columns

film_data = film_data[['title', 'year', 'length', 'worldwide_revenue', 'domestic_revenue', 'domestic_percent', 'foreign_revenue', 'foreign_percent', 'number_ratings', 'average_rating', 'genres']]

print("Re-ordered Columns")
print("-"*50)
display(film_data.head(2))

## check for duplicates in the final data set 

duplicates_2 = film_data[film_data.duplicated(keep=False)]

print(f"total records: {len(film_data)}")
print(f"duplicate records: {len(duplicates_2[duplicates_2.duplicated(keep="first")])}")

## Check data types 

print("Final Data Types")
print("-"*50)
print(film_data.dtypes)

## Save the clean & fully merged data frame to a CSV file 

film_data.to_csv("2017_2019_final_film_data.csv", header=True, index=False, encoding='utf-8')
